+++ libs/src/ui/element/Dialog/Dialog.stories.js
@@ -75,7 +75,7 @@ export const DialogDefault = {
 	args: {
 		hasClose: true,
 		className: '',
-		header: `<Title themeColor="primary" :level="2" >Title</Title>`,
+		header: `<Title :level="2" >Title</Title>`,
 		content: `<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard</p>`,
 		footer: `<Grid><Row :hasGap="true"><Column :sm="4"><Button variant="filled" radius="md" width="fluid" @click="onConfirm">OK</Button></Column><Column :sm="4"><Button variant="soft" radius="md" width="fluid" @click="onCancel">Cancel</Button></Column></Row></Grid>`,
 	},
@@ -109,7 +109,7 @@ export const DialogDefault = {
 					<Icon name="close" size="20"></Icon>
 				</button>
 				<div class="ded-dialog-header">
-					<Title themeColor="primary" :level="2" >Title</Title>
+					<Title :level="2" >Title</Title>
 				</div>
 				<div class="ded-dialog-body">
 					<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard</p>
@@ -148,7 +148,7 @@ export const DialogDefault = {
 						`  className=""`,
 						`>`,
 						`  <template #header>`,
-						`    <Title themeColor="primary" :level="2">Title</Title>`,
+						`    <Title :level="2">Title</Title>`,
 						`  </template>`,
 						`  <template #content>`,
 						`    <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard</p>`,
@@ -224,7 +224,7 @@ export const DialogDemo = {
 				className=""
 			>
 				<template #header>
-					<Title themeColor="primary" :level="2" >Title</Title>
+					<Title :level="2" >Title</Title>
 				</template>
 				<template #content>
 					<p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard</p>
@@ -270,7 +270,7 @@ export const DialogDemo = {
 						`  className=""`,
 						`>`,
 						`  <template #header>`,
-						`    <Title themeColor="primary" :level="2">Title</Title>`,
+						`    <Title :level="2">Title</Title>`,
 						`  </template>`,
 						`  <template #content>`,
 						`    <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard</p>`,