+++ libs/src/ui/element/Progress/LineProgress/LineProgress.stories.js
@@ -5,24 +5,24 @@ export default {
     component: LineProgress,
     tags: ["autodocs"],
     argTypes: {
-        themeColor: {
-            description: "主題顏色",
-            control: { type: "select" },
-            options: [
-                "primary",
-                "secondary",
-                "neutral",
-                "info",
-                "success",
-                "warning",
-                "error",
-            ],
-            table: {
-                type: {
-                    summary: "primary | secondary | neutral | info | success | warning | error"
-                }
-            }
-        },
+        // themeColor: {
+        //     description: "主題顏色",
+        //     control: { type: "select" },
+        //     options: [
+        //         "primary",
+        //         "secondary",
+        //         "neutral",
+        //         "info",
+        //         "success",
+        //         "warning",
+        //         "error",
+        //     ],
+        //     table: {
+        //         type: {
+        //             summary: "primary | secondary | neutral | info | success | warning | error"
+        //         }
+        //     }
+        // },
         label: {
             description: "進度指示標籤",
             control: { type: "text" },
@@ -64,7 +64,6 @@ export default {
 export const DefaultLineProgress = {
     name: "預設項目",
     args: {
-        themeColor: "primary",
         label: "Label",
         percent: 65,
         strokeWidth: 10,
@@ -79,8 +78,7 @@ export const DefaultLineProgress = {
             };
         },
         template: `
-            <LineProgress 
-                :themeColor="args.themeColor"
+            <LineProgress
                 :label="args.label"
                 :percent="args.percent"
                 :strokeWidth="args.strokeWidth"
@@ -99,7 +97,6 @@ export const DefaultLineProgress = {
                     const { args } = storyContext;
                     return [
                         '<LineProgress',
-                        `  :themeColor="${args.themeColor}"`,
                         `  :label="${args.label}"`,
                         `  :percent="${args.percent}"`,
                         `  :strokeWidth="${args.strokeWidth}"`,
@@ -116,7 +113,6 @@ export const DefaultLineProgress = {
 export const LineProgressLabelStory = {
     name: "顯示標籤",
     args: {
-        themeColor: "primary",
         label: "Label",
         percent: 65,
         strokeWidth: 10,
@@ -131,8 +127,7 @@ export const LineProgressLabelStory = {
             };
         },
         template: `
-            <LineProgress 
-                :themeColor="args.themeColor"
+            <LineProgress
                 :label="args.label"
                 :percent="args.percent"
                 :strokeWidth="args.strokeWidth"
@@ -151,7 +146,6 @@ export const LineProgressLabelStory = {
                     const { args } = storyContext;
                     return [
                         '<LineProgress',
-                        `  :themeColor="${args.themeColor}"`,
                         `  :label="${args.label}"`,
                         `  :percent="${args.percent}"`,
                         `  :strokeWidth="${args.strokeWidth}"`,
@@ -166,137 +160,137 @@ export const LineProgressLabelStory = {
 };
 
 //==== 主題色彩 ====//
-export const LineProgressColorStory = {
-    name: "主題色彩",
-    args: {
-        label: "",
-        strokeWidth: 10,
-        className: '',
-    },
-    render: (args) => ({
-        components: { LineProgress },
-        setup() {
-            // Create a ref for modelValue to be used with v-model
-            return {
-                args,
-            };
-        },
-        template: `
-            <div style="display:flex; flex-direction: column; gap:8px">
-                <LineProgress
-                    themeColor="primary"
-                    :label="args.label"
-                    :percent=40
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-                <LineProgress
-                    themeColor="secondary"
-                    :label="args.label"
-                    :percent=50
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-                <LineProgress
-                    themeColor="neutral"
-                    :label="args.label"
-                    :percent=60
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-                <LineProgress
-                    themeColor="info"
-                    :label="args.label"
-                    :percent=70
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-                <LineProgress
-                    themeColor="success"
-                    :label="args.label"
-                    :percent=80
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-                <LineProgress
-                    themeColor="warning"
-                    :label="args.label"
-                    :percent=90
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-                <LineProgress
-                    themeColor="error"
-                    :label="args.label"
-                    :percent=100
-                    :strokeWidth="args.strokeWidth"
-                    :className="args.className"
-                ></LineProgress>
-            </div>
-        `,
-    }),
-    // 控制 controls 中能控制的參數
-    parameters: {
-        controls: {
-            exclude: ['themeColor', 'percent' ],
-        },
-        docs: {
-            source: {
-                transform: (src, storyContext) => {
-                    const { args } = storyContext;
-                    return [
-                        '<LineProgress',
-                        `  themeColor="primary"`,
-                        `  label="${args.label}"`,
-                        `  :percent="40"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                        '<LineProgress',
-                        `  themeColor="secondary"`,
-                        `  label="${args.label}"`,
-                        `  :percent="50"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                        '<LineProgress',
-                        `  themeColor="tertiary"`,
-                        `  label="${args.label}"`,
-                        `  :percent="60"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                        '<LineProgress',
-                        `  themeColor="success"`,
-                        `  label="${args.label}"`,
-                        `  :percent="70"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                        '<LineProgress',
-                        `  themeColor="warning"`,
-                        `  :percent="80"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  label="${args.label}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                        '<LineProgress',
-                        `  themeColor="error"`,
-                        `  label="${args.label}"`,
-                        `  :percent="90"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                        '<LineProgress',
-                        `  themeColor="info"`,
-                        `  label="${args.label}"`,
-                        `  :percent="100"`,
-                        `  :strokeWidth="${args.strokeWidth}"`,
-                        `  className="${args.className}"`,
-                        '></LineProgress>',
-                    ].join('\n').trim();
-                }
-            }
-        }
-    },
-};
+// export const LineProgressColorStory = {
+//     name: "主題色彩",
+//     args: {
+//         label: "",
+//         strokeWidth: 10,
+//         className: '',
+//     },
+//     render: (args) => ({
+//         components: { LineProgress },
+//         setup() {
+//             // Create a ref for modelValue to be used with v-model
+//             return {
+//                 args,
+//             };
+//         },
+//         template: `
+//             <div style="display:flex; flex-direction: column; gap:8px">
+//                 <LineProgress
+//                     themeColor="primary"
+//                     :label="args.label"
+//                     :percent=40
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//                 <LineProgress
+//                     themeColor="secondary"
+//                     :label="args.label"
+//                     :percent=50
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//                 <LineProgress
+//                     themeColor="neutral"
+//                     :label="args.label"
+//                     :percent=60
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//                 <LineProgress
+//                     themeColor="info"
+//                     :label="args.label"
+//                     :percent=70
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//                 <LineProgress
+//                     themeColor="success"
+//                     :label="args.label"
+//                     :percent=80
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//                 <LineProgress
+//                     themeColor="warning"
+//                     :label="args.label"
+//                     :percent=90
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//                 <LineProgress
+//                     themeColor="error"
+//                     :label="args.label"
+//                     :percent=100
+//                     :strokeWidth="args.strokeWidth"
+//                     :className="args.className"
+//                 ></LineProgress>
+//             </div>
+//         `,
+//     }),
+//     // 控制 controls 中能控制的參數
+//     parameters: {
+//         controls: {
+//             exclude: ['themeColor', 'percent' ],
+//         },
+//         docs: {
+//             source: {
+//                 transform: (src, storyContext) => {
+//                     const { args } = storyContext;
+//                     return [
+//                         '<LineProgress',
+//                         `  themeColor="primary"`,
+//                         `  label="${args.label}"`,
+//                         `  :percent="40"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                         '<LineProgress',
+//                         `  themeColor="secondary"`,
+//                         `  label="${args.label}"`,
+//                         `  :percent="50"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                         '<LineProgress',
+//                         `  themeColor="tertiary"`,
+//                         `  label="${args.label}"`,
+//                         `  :percent="60"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                         '<LineProgress',
+//                         `  themeColor="success"`,
+//                         `  label="${args.label}"`,
+//                         `  :percent="70"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                         '<LineProgress',
+//                         `  themeColor="warning"`,
+//                         `  :percent="80"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  label="${args.label}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                         '<LineProgress',
+//                         `  themeColor="error"`,
+//                         `  label="${args.label}"`,
+//                         `  :percent="90"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                         '<LineProgress',
+//                         `  themeColor="info"`,
+//                         `  label="${args.label}"`,
+//                         `  :percent="100"`,
+//                         `  :strokeWidth="${args.strokeWidth}"`,
+//                         `  className="${args.className}"`,
+//                         '></LineProgress>',
+//                     ].join('\n').trim();
+//                 }
+//             }
+//         }
+//     },
+// };